【作業中】ユーザライブラリにNewlibを使用してみる
いろいろ変えてビルドはできるようになったが（mkfs以外）、
ファイルサイズがでかくなりfsの上限に引っかかってしまい、
shが組み込めない。

Newlibのconfigure: [Newlibビルドメモ - uchan note](http://uchan.hateblo.jp/entry/2017/07/24/234030)
CC=gcc ../newlib/newlib/configure --disable-multilib

XV6をビルドしているものとと同じオプションでmakeさせるため、
Makefile をいじる。

+++ user/sh.c
@@ -455,7 +455,7 @@ backcmd(struct cmd *subcmd)
 static const char symbols[] = "<|>&;()#";
 
 char *skipws(char *s, char *es) {
-  while(s < es && isspace(*s))
+  while(s < es && isspace((int)*s))
     s++;
   return s;
 }
@@ -541,7 +541,7 @@ gettoken(char **ps, char *es, char **q)
       }
       break;
     default:
-      while(s < es && !isspace(*s) && strchr(symbols, *s) == NULL)
+      while(s < es && !isspace((int)*s) && strchr(symbols, *s) == NULL)
         s++;
       break;
     }
